Renamed the google-oauth and github-oauth folder in the modules directory and updated all import paths where the folder is used.
Renamed the google-oauth and github-oauth folder in the modules directory and updated all import paths where the folder is used.
