+++ modules/github-oauth/init.js
